Rename todo_repository to citation_repository
