+++ src/app.py
@@ -1,7 +1,7 @@
 import datetime
 from flask import redirect, render_template, request, jsonify, flash, url_for, Response, abort
 from db_helper import reset_db
-from repositories.todo_repository import (
+from repositories.citation_repository import (
     get_cites, check_citation_type, get_cite_by_id, update_citation,
     create_book_citation, create_article_citation, create_inproceedings_citation,
     remove_citation1
